// src/shared/constants/error-codes.enum.ts
export enum ErrorCode {
  // Common
  FAILED_TO_UPLOAD_IMAGE = 'FAILED_TO_UPLOAD_IMAGE',

  // Auth
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_INVALID_SITE = 'AUTH_INVALID_SITE',
  AUTH_EMAIL_ALREADY_EXISTS = 'AUTH_EMAIL_ALREADY_EXISTS',
  AUTH_USERNAME_ALREADY_EXISTS = 'AUTH_USERNAME_ALREADY_EXISTS',
  AUTH_PHONE_NUMBER_ALREADY_EXISTS = 'AUTH_PHONE_NUMBER_ALREADY_EXISTS',
  AUTH_CANNOT_CREATE_USER = 'AUTH_CANNOT_CREATE_USER',
  AUTH_CANNOT_SEND_VERIFICATION_EMAIL = 'AUTH_CANNOT_SEND_VERIFICATION_EMAIL',
  AUTH_INVALID_VERIFICATION_CODE = 'AUTH_INVALID_VERIFICATION_CODE',
  AUTH_USER_NOT_FOUND = 'AUTH_USER_NOT_FOUND',
  AUTH_EMAIL_ALREADY_VERIFIED = 'AUTH_EMAIL_ALREADY_VERIFIED',
  AUTH_INVALID_REFRESH_TOKEN = 'AUTH_INVALID_REFRESH_TOKEN',
  AUTH_EMAIL_NOT_VERIFIED = 'AUTH_EMAIL_NOT_VERIFIED',
  AUTH_INVALID_PWD_RESET_CODE = 'AUTH_INVALID_PWD_RESET_CODE',
  AUTH_USER_LOCKED = 'AUTH_USER_LOCKED',
  AUTH_IS_NOT_ADMIN = 'AUTH_IS_NOT_ADMIN',
  AUTH_USER_NOT_VERIFIED = 'AUTH_USER_NOT_VERIFIED',
  AUTH_PASSWORD_CONFIRM_MISMATCH = 'AUTH_PASSWORD_CONFIRM_MISMATCH',
  AUTH_PASSWORD_INVALID_CURRENT = 'AUTH_PASSWORD_INVALID_CURRENT',
  AUTH_PASSWORD_REUSED = 'AUTH_PASSWORD_REUSED',

  // Social Auth Error Codes
  AUTH_SOCIAL_INVALID_GOOGLE_TOKEN = 'AUTH_SOCIAL_INVALID_GOOGLE_TOKEN',
  AUTH_SOCIAL_INVALID_FACEBOOK_TOKEN = 'AUTH_SOCIAL_INVALID_FACEBOOK_TOKEN',
  AUTH_SOCIAL_EMAIL_ALREADY_EXISTS = 'AUTH_SOCIAL_EMAIL_ALREADY_EXISTS',
  AUTH_SOCIAL_PROVIDER_NOT_LINKED = 'AUTH_SOCIAL_PROVIDER_NOT_LINKED',
  AUTH_SOCIAL_NOT_SITE_USER = 'AUTH_SOCIAL_NOT_SITE_USER',
  AUTH_SOCIAL_EMAIL_REQUIRED = 'AUTH_SOCIAL_EMAIL_REQUIRED',
  AUTH_UPDATE_PROFILE_FORBIDDEN = 'AUTH_UPDATE_PROFILE_FORBIDDEN',

  SHOP_REQUEST_HAS_EXISTED = 'SHOP_REQUEST_HAS_EXISTED',
  SHOP_REQUEST_NOT_FOUND = 'SHOP_REQUEST_NOT_FOUND',
  SHOP_REQUEST_NOT_APPROVED = 'SHOP_REQUEST_NOT_APPROVED',
  SHOP_REQUEST_FAILED = 'SHOP_REQUEST_FAILED',
  SHOP_HAS_EXISTED = 'SHOP_HAS_EXISTED',
  SHOP_NOT_FOUND = 'SHOP_NOT_FOUND',
  SHOP_INFORMATION_NOT_FOUND = 'SHOP_INFORMATION_NOT_FOUND',
  SHOP_SUSPENDED = 'SHOP_SUSPENDED',
  SHOP_USER_DOES_NOT_HAVE_A_SHOP = 'SHOP_USER_DOES_NOT_HAVE_A_SHOP',
  SHOP_PRODUCT_INVALID_CATEGORY_AND_SUBCATEGORY = 'SHOP_PRODUCT_INVALID_CATEGORY_AND_SUBCATEGORY',
  SHOP_PRODUCT_NOT_FOUND = 'SHOP_PRODUCT_NOT_FOUND',
  SHOP_PRODUCT_VERSION_NOT_FOUND = 'SHOP_PRODUCT_VERSION_NOT_FOUND',
  SHOP_PRODUCT_VERSION_CANNOT_APPROVE_OR_REJECT = 'SHOP_PRODUCT_VERSION_CANNOT_APPROVE_OR_REJECT',
  SHOP_PRODUCT_CANNOT_UPDATE_SUSPENDED = 'SHOP_PRODUCT_CANNOT_UPDATE_SUSPENDED',
  SHOP_PRODUCT_INVALID_REJECTED_REASON = 'SHOP_PRODUCT_INVALID_REJECTED_REASON',
  SHOP_PRODUCT_STATE_CHANGE_NOT_ALLOWED = 'SHOP_PRODUCT_STATE_CHANGE_NOT_ALLOWED',
  SHOP_PRODUCT_INVALID_STATE_TRANSITION = 'SHOP_PRODUCT_INVALID_STATE_TRANSITION',
  SHOP_SUSPENSION_REASON_REQUIRED = 'SHOP_SUSPENSION_REASON_REQUIRED',

  PRODUCT_BOOST_NOT_OWNER = 'PRODUCT_BOOST_NOT_OWNER',
  PRODUCT_BOOST_ALREADY_ACTIVE = 'PRODUCT_BOOST_ALREADY_ACTIVE',
  PRODUCT_BOOST_PACKAGE_MISMATCH = 'PRODUCT_BOOST_PACKAGE_MISMATCH',
  PRODUCT_BOOST_PACKAGE_NOT_FOUND = 'PRODUCT_BOOST_PACKAGE_NOT_FOUND',
  PRODUCT_BOOST_PACKAGE_VERSION_INVALID = 'PRODUCT_BOOST_PACKAGE_VERSION_INVALID',
  PRODUCT_BOOST_PACKAGE_VERSION_MISSING = 'PRODUCT_BOOST_PACKAGE_VERSION_MISSING',
  PRODUCT_BOOST_WALLET_LOCKED = 'PRODUCT_BOOST_WALLET_LOCKED',
  PRODUCT_BOOST_INSUFFICIENT_BALANCE = 'PRODUCT_BOOST_INSUFFICIENT_BALANCE',
  PRODUCT_BOOST_PRODUCT_IS_NOT_LIVE = 'PRODUCT_BOOST_PRODUCT_IS_NOT_LIVE',

  // Cart
  CART_PRODUCT_ALREADY_EXISTS = 'CART_PRODUCT_ALREADY_EXISTS',

  // Wallet
  WALLET_NOT_FOUND = 'WALLET_NOT_FOUND',
  WALLET_INSUFFICIENT_BALANCE = 'WALLET_INSUFFICIENT_BALANCE',
  WALLET_IS_LOCKED = 'WALLET_IS_LOCKED',
  WALLET_WITHDRAW_APPROVE_FAILED = 'WALLET_WITHDRAW_APPROVE_FAILED',
  WALLET_TRANSACTION_NOT_FOUND = 'WALLET_TRANSACTION_NOT_FOUND',
  WALLET_TRANSACTION_NOT_PENDING = 'WALLET_TRANSACTION_NOT_PENDING',

  // Order error codes
  ORDER_PRODUCT_OR_VERSION_NOT_FOUND = 'ORDER_PRODUCT_OR_VERSION_NOT_FOUND',
  ORDER_SHOP_OWNER_USER_NOT_FOUND = 'ORDER_SHOP_OWNER_USER_NOT_FOUND',
  ORDER_FAILED_TO_CREATE_ORDER = 'ORDER_FAILED_TO_CREATE_ORDER',
  ORDER_PRODUCT_VERSION_ALREADY_PURCHASED = 'ORDER_PRODUCT_VERSION_ALREADY_PURCHASED',

  PAGINATION_INVALID_PAGE = 'PAGINATION_INVALID_PAGE',
  PAGINATION_INVALID_LIMIT = 'PAGINATION_INVALID_LIMIT',

  PREVIEW_OFF_THIS_USER_IS_EXIST = 'PREVIEW_OFF_THIS_USER_IS_EXIST',
}